requirements.txt
networkx==3.1
python-constraint==1.4.0
python-sat==0.1.8.dev10
pyperplan==2.1
numpy==1.26.0
scipy==1.11.3
prettytable==3.9.0

# The following packages are necessary for machine learning assignments but they may not be installed on recodex.
scikit-learn==1.3.2
pandas==2.2.0
tensorflow==2.15.0
keras==2.15.0
matplotlib==3.8.3

